requirements.txt
pika==1.3.2  # RabbitMQ client
aiogram==3.4.1  # Telegram Bot API
fastapi==0.111.0
uvicorn==0.30.1
asyncpg==0.29.0  # Для хранения chat_id
SQLAlchemy==2.0.30
python-dotenv==1.0.1
pydantic
pydantic_settings
pydantic-core
greenlet
black
httpx
python-json-logger
python-dateutil